Add files via upload
This is the first rough draft with skeleton code for classes and functions. We have commented out several areas where we still have work to do. These sections include determining whether the Doctors and Nurses should be objects or just a vector, determining whether a doctor or nurse is available to treat a patient, storing patient visits, and creating patient names. We have work to do determining which data structures will be best in each situation. We have yet to implement any unit tests.