Add files via upload
Several changes. Function definitions in multiple classes altered. Polymorphism shifted from Caregiver to Doctor and Nurse to EmergencyRoom to Doctor and Nurse. Queue of Doctor objects and queue of nurse objects created in place of the vectors of EmergencyRoom objects. Functions moved from EmergencyRoom class to Doctor and Nurse classes.