Add files via upload
In this version, we removed a lot of excess, unnecessary code and organizes our code in a way that helped us understand it better. We renamed several variables, queues, and vectors so our code was more understandable as well. We reworked the Simulator, WaitingRoom, EmergencyRoom, and Discharge queues and feel as though they are almost complete. The Doctor and Nurse queues were changed into vectors and we now have only one priority queue instead of three. We still have to work on polymorphism in the Caregiver, Doctor, and Nurse classes, the display menu, unit testing, and retrieving and assigning patient names.