Add files via upload
In this version, we have done work on the relationship between the EmergencyRoom and the Caregiver class. We have also worked  with the contents of the Doctor and Nurse classes. We decided to move a lot of the code from the update function in Doctor.h and Nurses.h to EmergencyRoom. The goal of our next version is to finish implementing the polymorphism.